+++ back-end/src/main/java/com/ring/bookstore/dtos/BookDTO.java
@@ -2,13 +2,13 @@ package com.ring.bookstore.dtos;
 
 import java.math.BigDecimal;
 
-public record BookDTO(Integer id,
+public record BookDTO(Long id,
                       String title,
                       String description,
                       String image,
                       Double price,
                       BigDecimal onSale,
-                      Integer amount,
+                      Short amount,
                       Double rating,
                       Integer orderTime) {
