+++ back-end/src/main/java/com/ring/bookstore/service/OrderService.java
@@ -13,9 +13,9 @@ import com.ring.bookstore.request.OrderRequest;
 public interface OrderService {
 	
 	Page<OrderDTO> getAllOrders(Account user, Integer pageNo, Integer pageSize, String sortBy, String sortDir);
-	Page<OrderDTO> getOrdersByBookId(Integer id, Integer pageNo, Integer pageSize, String sortBy, String sortDir);
+	Page<OrderDTO> getOrdersByBookId(Long id, Integer pageNo, Integer pageSize, String sortBy, String sortDir);
 	Page<OrderDTO> getOrdersByUser(Account user, Integer pageNo, Integer pageSize);
-	OrderDTO getOrderById(Integer id);
+	OrderDTO getOrderById(Long id);
 	public OrderReceipt checkout(OrderRequest request, Account user);
 	List<ChartDTO> getMonthlySale(Account user);
 }
